Calculate Metabolic Rate

This function compute metabolic rate (Cal/m2/hr) from Harvard Spirometer tracing, subject's weight and height, and environmental condition i.e., barometric pressure and temperature. The output printed to console will show each step of calculation which is suitable for educational purposes.

get_metabolic_rate(
  x,
  y,
  paper_speed = 25,
  baro,
  temp_c,
  wt_kg,
  ht_cm,
  cal_eqi_oxygen = 4.825
)

Arguments

x
Displacement of tracing in x-direction (in millimeter).

y
Displacement of tracing in y-direction (in millimeter).

paper_speed
Paper speed of the kymograph (in millimeter/minute), 25 ml/min by default.

baro
(If condition = "STPD") Barometric pressure at the recording site.

temp_c
(If condition = "STPD") Temperature in celsius at the recording site.

wt_kg
(Numeric) Weight of the subject in kilogram

ht_cm
(Numeric) Height of the subject in centimetre

cal_eqi_oxygen
(Numeric) Caloric equivalent of Oxygen (default is 4.825, at RQ = 0.82)

Value

Report printed to console, return metabolic rate in Cal/m2/hr, invisibly

Examples

get_metabolic_rate(x = 20, y = 10,
                   baro = 760, temp_c = 25,
                   wt_kg = 80, ht_cm = 180)
#> Harvard spirometer tracing:
#>  - Paper speed = 25 mm/min 
#>  - Time interval = 0.8 min (horizontal displacement = 20 mm)
#>  - Volume change = 300 ml (vertical displacement = 10 mm)
#> 
#> Oxygen Consumption at ATPS = 22.5 L/hr
#> 
#> Metabolic rate calculation:
#>  - STPD correction factor = 0.883 (760 mmHg, 25 degree celcius) 
#>  - Oxygen Consumption at STPD = 19.869 L/hr (0.883 x 22.5 L/hr)
#>  - Caloric equivalent of Oxygen = 4.825 Cal/L of Oxygen
#>  - BSA = 1.996 square metre (wt = 80 kg, ht = 180 cm)
#> 
#> Metabolic Rate = 48.02 Cal/m2/hr 
#> 
#>